test(app): share workspace slug wait helper across e2e specs
